fixing slit as well
