+++ nexus_constructor/add_component_window.py
@@ -1,4 +1,3 @@
-import logging
 from collections import OrderedDict
 from copy import deepcopy
 from functools import partial
@@ -8,26 +7,13 @@ from PySide2.QtCore import Qt, QUrl, Signal
 from PySide2.QtGui import QKeyEvent, QVector3D
 from PySide2.QtWidgets import QListWidget, QListWidgetItem, QMessageBox, QWidget
 
-from nexus_constructor.common_attrs import (
-    NX_CLASSES_WITH_PLACEHOLDERS,
-    SHAPE_GROUP_NAME,
-    CommonAttrs,
-)
+from nexus_constructor.common_attrs import NX_CLASSES_WITH_PLACEHOLDERS, CommonAttrs
 from nexus_constructor.component_tree_model import NexusTreeModel
-from nexus_constructor.component_type import (
-    COMPONENT_TYPES,
-    PIXEL_COMPONENT_TYPES,
-    SLIT_CLASS_NAME,
-)
+from nexus_constructor.component_type import COMPONENT_TYPES, PIXEL_COMPONENT_TYPES
 from nexus_constructor.field_utils import get_fields_with_update_functions
 from nexus_constructor.field_widget import FieldWidget
-from nexus_constructor.geometry.disk_chopper.disk_chopper_checker import ChopperChecker
-from nexus_constructor.geometry.disk_chopper.disk_chopper_geometry_creator import (
-    DiskChopperGeometryCreator,
-)
 from nexus_constructor.geometry.geometry_loader import load_geometry
 from nexus_constructor.geometry.pixel_data import PixelData, PixelGrid, PixelMapping
-from nexus_constructor.geometry.slit.slit_geometry import SlitGeometry
 from nexus_constructor.model import Group, GroupContainer
 from nexus_constructor.model.component import Component
 from nexus_constructor.model.geometry import (
@@ -54,25 +40,6 @@ from nexus_constructor.validators import (
     UnitValidator,
 )
 from ui.add_component import Ui_AddComponentDialog
-
-
-def _set_chopper_geometry(component: Component):
-    """
-    Attempts to set a chopper geometry in the component by checking if the component fields describe a valid chopper.
-    :param component: The component to be given a shape.
-    """
-    chopper_validator = ChopperChecker(component.children)
-    if chopper_validator.validate_chopper():
-        chopper_details = chopper_validator.chopper_details
-        chopper_creator = DiskChopperGeometryCreator(chopper_details)
-        component[SHAPE_GROUP_NAME] = chopper_creator.create_disk_chopper_geometry()
-    else:
-        logging.warning("Validation failed. Unable to create disk chopper mesh.")
-
-
-def _set_slit_geometry(component: Component):
-    slit_geometry = SlitGeometry(component)
-    component.set_off_shape(slit_geometry.create_slit_geometry())
 
 
 class AddComponentDialog(Ui_AddComponentDialog):
@@ -540,11 +507,6 @@ class AddComponentDialog(Ui_AddComponentDialog):
                 filename=self.fileLineEdit.text(),
                 pixel_data=pixel_data,
             )
-        elif (
-            self.noShapeRadioButton.isChecked()
-            and component.nx_class == SLIT_CLASS_NAME
-        ):
-            _set_slit_geometry(component)
 
     def get_pixel_visibility_condition(self) -> bool:
         """
